add chunking by components, add alias to relative imports, change login logic to fast page load and preload chunked protectedApp component
