+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/package.json
@@ -7,7 +7,7 @@
     "test": "NODE_ENV=test NODE_TLS_REJECT_UNAUTHORIZED='0' mocha --recursive --require babel-register  -r mock-local-storage \"./source/test/**/*.test.js\"",
     "build:prod": "NODE_ENV=production NODE_OPTIONS=--max_old_space_size=4096 webpack -p --display errors-only --hide-modules",
     "build:dev": "NODE_ENV=development webpack -d --watch ",
-    "analysis": "NODE_ENV=development webpack --env.analysis -d --progress",
+    "analysis": "NODE_ENV=development webpack --env.analysis -p --progress",
     "lint": "eslint source"
   },
   "repository": {
@@ -37,6 +37,7 @@
     "react-dom": "^16.3.0",
     "react-dropzone": "^3.13.4",
     "react-intl": "^2.4.0",
+    "react-loadable": "^5.5.0",
     "react-modal": "^2.2.2",
     "react-quill": "1.0.0",
     "react-router-dom": "^4.2.2",
@@ -52,7 +53,7 @@
     "babel-core": "^6.24.1",
     "babel-eslint": "^8.2.2",
     "babel-loader": "^7.0.0",
-    "babel-plugin-dynamic-import-webpack": "^1.0.1",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.6.1",
