+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Details/Overview/BusinessPlans.jsx
@@ -4,9 +4,10 @@ import ListItem from '@material-ui/core/ListItem';
 import ListItemText from '@material-ui/core/ListItemText';
 import PropTypes from 'prop-types';
 
+import API from 'AppData/api';
+import { Progress } from 'AppComponents/Shared';
+
 import APIPropertyField from './APIPropertyField';
-import API from '../../../../data/api';
-import { Progress } from '../../../Shared';
 
 /**
  *
