+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Details/components/APIDetailsRoutes.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
 import PropTypes from 'prop-types';
+import Utils from 'AppData/Utils';
 
 import { PageNotFound } from '../../../Base/Errors/';
 import EnvironmentOverview from '../EnvironmentOverview/EnvironmentOverview';
-import Utils from '../../../../data/Utils';
 // import PermissionFormWrapper from '../Permissions/Permission';
 
 /**
