+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Details/components/DeleteApiButton.jsx
@@ -6,9 +6,9 @@ import Slide from '@material-ui/core/Slide';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 
-import { resourceMethod, resourcePath, ScopeValidation } from '../../../../data/ScopeValidation';
-import ApiPermissionValidation from '../../../../data/ApiPermissionValidation';
-import Alert from '../../../Shared/Alert';
+import { resourceMethod, resourcePath, ScopeValidation } from 'AppData/ScopeValidation';
+import ApiPermissionValidation from 'AppData/ApiPermissionValidation';
+import Alert from 'AppComponents/Shared/Alert';
 
 /**
  * Handle Delete an API from API Overview/Details page
